Create new file for functional grasp metrics, move ContactGrasp metric to that file
